+++ CHANGELOG.md
@@ -16,6 +16,13 @@ Types of changes
 
 # Latch SDK Changelog
 
+## 2.58.0 - 2024-03-26
+
+### Fixed
+
+* `latch launch` error messages being suppressed
+* `g6e-12xlarge` GPU instance type incorrect GPU annotation (1 -> 4 L40s GPUs)
+
 ## 2.57.3 - 2024-03-19
 
 ### Fixed